Allow static calls to API methods
