+++ includes/class-wp-typography-factory.php
@@ -2,7 +2,7 @@
 /**
  *  This file is part of wp-Typography.
  *
- *  Copyright 2017 Peter Putzer.
+ *  Copyright 2017-2018 Peter Putzer.
  *
  *  This program is free software; you can redistribute it and/or
  *  modify it under the terms of the GNU General Public License
@@ -71,6 +71,13 @@ abstract class WP_Typography_Factory {
 			] );
 			self::$factory->addRule( Options::class, [
 				'shared' => true,
+			] );
+
+			// API implementation.
+			self::$factory->addRule( 'substitutions', [
+				\WP_Typography::class => [
+					'instance' => \WP_Typography\Implementation::class,
+				],
 			] );
 
 			// Load version from plugin data.
